SC_ui/requirements.txt
pandas==1.0.5
numpy==1.19.0
streamlit==0.72.0
Pillow==8.2.0
